refactor(knowledge): extract build/sync services, thin router (1132→~280 lines)
Move folder sync and knowledge-base build orchestration out of
routers/knowledge.py into dedicated service modules. Router now only
does parameter parsing + DI + response marshalling + SSE/WS forwarding.

New modules:
- app/services/knowledge_build/sync_service.py — KnowledgeSyncService
  absorbs _sync_folder (was 285 lines in router): fetches B站 videos,
  upserts collection rows, handles empty-response protection.
- app/services/knowledge_build/build_service.py — KnowledgeBuildService
  absorbs _build_knowledge_base_task (was 192 lines): cursor tracking,
  legacy in-memory state, WS broadcast, per-folder vectorization.
  Process-wide singleton lock (_active_build_task_id) moved here.
- app/repository/knowledge_repository.py — query endpoints' DB reads:
  get_folder_status, get_vectorized_pages.

Router changes:
- 835 lines removed; 4 query endpoints delegate to repository
- 2 sync/build endpoints delegate to services
- Background task strong-ref set moved to service
- SSE/WS event forwarding stays in router (HTTP-layer concern)

Why: CLAUDE.md §6 forbids DB operations in routers — knowledge.py had
20 inline db.execute/add/commit calls and was 1132 lines, effectively
a service with route decorators. The build task's WS broadcast + lock
was impossible to unit test in the router layer.

Risk notes:
- _build_knowledge_base_task contains legacy in-memory state migration
  logic — behavior preserved verbatim, just relocated.
- Process-wide singleton lock semantics unchanged (still per-process,
  not per-user; acceptable for single-worker deployment).
