+++ app/services/knowledge_build/__init__.py
@@ -1,0 +1,15 @@
+"""Knowledge build services — sync + build orchestration.
+
+Split out of ``app/routers/knowledge.py`` to keep the router a thin HTTP layer.
+"""
+from app.services.knowledge_build.sync_service import KnowledgeSyncService
+from app.services.knowledge_build.build_service import (
+    KnowledgeBuildService,
+    get_build_service,
+)
+
+__all__ = [
+    "KnowledgeSyncService",
+    "KnowledgeBuildService",
+    "get_build_service",
+]
